requirements.txt
aiogram==3.1.1
Flask==3.0.0
Flask-Login==0.6.3
psutil==5.9.6
python-dotenv==1.0.0
Werkzeug==3.0.1
